requirements.txt
# Here all the packages required to run the project are listed with their versions.
# We need to read and install these packages before running the project.
# ---------------------------------------------------------------------------------

Django==5.1
django-autoslug==1.9.9
django-dirtyfields==1.9.2
django-phonenumber-field==7.0.2
django-redis==5.3.0
django-silk==5.0.3
django-simple-history==3.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3
drf-spectacular==0.26.3
phonenumbers==8.13.10
django-filter==23.1
